Move init into local-cli
